+++ src/main/java/eterea/migration/rest/configuration/WordPressClientConfig.java
@@ -14,21 +14,12 @@ public class WordPressClientConfig {
       this.environment = environment;
    }
 
-   @Bean(name = "wordPressApiClientSite1")
-   public WordPressApiClient wordPressApiClientSite1() {
-      String baseUrl = environment.getProperty("app.wordpress.site1.base-url");
-      String username = environment.getProperty("app.wordpress.site1.username");
-      String password = environment.getProperty("app.wordpress.site1.password");
-      String siteName = environment.getProperty("app.wordpress.site1.name", "site1");
-      return new WordPressApiClient(baseUrl, username, password, siteName);
-   }
-
-   @Bean(name = "wordPressApiClientSite2")
-   public WordPressApiClient wordPressApiClientSite2() {
-      String baseUrl = environment.getProperty("app.wordpress.site2.base-url");
-      String username = environment.getProperty("app.wordpress.site2.username");
-      String password = environment.getProperty("app.wordpress.site2.password");
-      String siteName = environment.getProperty("app.wordpress.site2.name", "site2");
+   @Bean(name = "wordPressApiClient")
+   public WordPressApiClient wordPressApiClient() {
+      String baseUrl = environment.getProperty("app.wordpress.site.base-url");
+      String username = environment.getProperty("app.wordpress.site.username");
+      String password = environment.getProperty("app.wordpress.site.password");
+      String siteName = environment.getProperty("app.wordpress.site.name", "site");
       return new WordPressApiClient(baseUrl, username, password, siteName);
    }
 }