+++ src/main/java/eterea/migration/rest/extern/OrderNoteWeb.java
@@ -1,5 +1,6 @@
 package eterea.migration.rest.extern;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,6 +12,7 @@ import java.util.List;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class OrderNoteWeb {
 
     @JsonProperty("order_number")